+++ react-components/src/bundle.js
@@ -1,0 +1,5 @@
+import CookiesModal from './CookiesModal'
+
+export default {
+  CookiesModal,
+};