+++ react-components/webpack.config.js
@@ -1,0 +1,27 @@
+const path = require('path');
+
+module.exports = {
+    entry: {
+        'reactComponents': './react-components/src/bundle.js',
+    },
+    output: {
+        path: path.resolve(__dirname, '../directory_components/static/directory_components/js/'),
+        filename: '[name].js',
+        publicPath: '',
+        library: '[name]',
+        libraryExport: 'default',
+        libraryTarget: 'var'
+    },
+    resolve: {
+        extensions: ['.js', '.jsx']
+    },
+    module: {
+        rules: [
+            {
+                test: /\.jsx?$/,
+                loader: 'babel-loader',
+                exclude: /node_modules/
+            }
+        ]
+    }
+};